+++ NEWS.md
@@ -1,5 +1,7 @@
 # Version 0.1.3
 * Added the `proximity` function to count the number samples within a radius
+* Removed the requirement for `add_xy` in the `benchmark()` function
+* Removed `filename` and `wopt` from all function and replaced by additional args `...`
 
 # Version 0.1.2
 * Check for availability of the `terra` package is added when input is a raster